Apply suggestions from code review

+++ ggml/src/ggml-metal/ggml-metal-impl.h
@@ -895,10 +895,10 @@ typedef struct {
 } ggml_metal_kargs_arange;
 
 typedef struct {
-    int64_t  ne00;
-    int64_t  ne01;
-    int64_t  ne02;
-    int64_t  ne03;
+    int32_t  ne00;
+    int32_t  ne01;
+    int32_t  ne02;
+    int32_t  ne03;
     uint64_t nb00;
     uint64_t nb01;
     uint64_t nb02;
